x/liquidity/types/tick.go
package types

import (
	"math"
	"math/big"

	sdk "github.com/cosmos/cosmos-sdk/types"
)

func PriceToTick(price sdk.Dec, prec int) sdk.Dec {
	b := price.BigInt()
	l := len(b.Text(10)) - 1 // floor(log10(b))
	d := int64(l - prec)
	if d > 0 {
		p := big.NewInt(10)
		p.Exp(p, big.NewInt(d), nil)
		b.Quo(b, p).Mul(b, p)
	}
	return sdk.NewDecFromBigIntWithPrec(b, sdk.Precision)
}

func TickToIndex(tick sdk.Dec, prec int) int {
	b := tick.BigInt()
	l := len(b.Text(10)) - 1 // floor(log10(b))
	d := int64(l - prec)
	if d > 0 {
		q := big.NewInt(10)
		q.Exp(q, big.NewInt(d), nil)
		b.Quo(b, q)
	}
	p := int(math.Pow10(prec))
	b.Sub(b, big.NewInt(int64(p)))
	return (l-prec)*9*p + int(b.Int64())
}

func TickFromIndex(i, prec int) sdk.Dec {
	p := int(math.Pow10(prec))
	l := i/(9*p) + prec
	t := big.NewInt(int64(p + i%(p*9)))
	if l > prec {
		m := big.NewInt(10)
		m.Exp(m, big.NewInt(int64(l-prec)), nil)
		t.Mul(t, m)
	}
	return sdk.NewDecFromBigIntWithPrec(t, sdk.Precision)
}
